requirements.txt
# Core AI & API
google-genai>=0.8.0
fastapi>=0.95.0
uvicorn>=0.21.0

# PC Control & Automation
pyautogui>=0.9.53
pygetwindow>=0.0.9
pillow>=9.0.0
pytesseract>=0.3.10
pywinauto>=0.6.8

# Configuration & Environment
python-dotenv>=0.21.0
pydantic>=1.9.0

# Database & Storage
sqlalchemy>=2.0.0
aiosqlite>=0.17.0

# HTTP & Async
httpx>=0.23.0
aiohttp>=3.8.0

# Cryptography & Security
cryptography>=41.0.0
PyJWT>=2.0.0

# Text-to-Speech
pyttsx3>=2.90

# Logging & Monitoring
python-json-logger>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0

# Optional: Image Recognition
opencv-python>=4.5.0

# Optional: PDF Processing
PyPDF2>=3.0.0

# Optional: Clipboard
pyperclip>=1.8.2

